+++ proyecto/src/main/java/com/Proyecto/service/CompraService.java
@@ -13,6 +13,6 @@ public interface CompraService {
     Compra save(Compra compra);
     String GenerarNumCompra();
     List<Compra> findByUsuarioId(Integer usuario);
-    Optional<Compra> findById(Integer id);
+    Optional<Compra> findById(Long id);
 
 }
